app/requirements.txt
black==24.4.2
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
Flask-Cors==4.0.1
greenlet==3.0.3
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3
